Move tail sampling config to its own package and remove stale configs
Follow up of #200. Second step of #146.
