--- service/builder/testdata/sampling_config.yaml
@@ -1,67 +1,0 @@
-receivers:
-  jaeger:
-    collector_http_port: 14268
-  opencensus:
-    address: "127.0.0.1:55678"
-queued-exporters:
-  jaeger1:
-    sender-type: jaeger-thrift-http
-    jaeger-thrift-http:
-      collector_endpoint: "http://host.docker.internal:14368/api/traces"
-  jaeger2:
-    sender-type: jaeger-thrift-http
-    jaeger-thrift-http:
-      collector_endpoint: "http://host.docker.internal:14468/api/traces"
-  jaeger3:
-    sender-type: jaeger-thrift-http
-    jaeger-thrift-http:
-      collector_endpoint: "http://host.docker.internal:14368/api/traces"
-  jaeger4:
-    sender-type: jaeger-thrift-http
-    jaeger-thrift-http:
-      collector_endpoint: "http://host.docker.internal:14468/api/traces"
-  jaeger5:
-    sender-type: jaeger-thrift-http
-    jaeger-thrift-http:
-      collector_endpoint: "http://host.docker.internal:14468/api/traces"
-sampling:
-  mode: tail
-  decision-wait: 31s
-  num-traces: 20001
-  expected-new-traces-per-sec: 100
-  policies:
-    string-attribute-filter1:
-        exporters:
-          - jaeger1
-        policy: string-attribute-filter
-        configuration:
-          key: "test"
-          values:
-            - "value 1"
-            - "value 2"
-    numeric-attribute-filter2:
-        exporters:
-          - jaeger2
-        policy: numeric-attribute-filter
-        configuration:
-          key: "http.status_code"
-          min-value: 400
-          max-value: 999
-    string-attribute-filter3:
-        exporters:
-          - jaeger3
-        policy: string-attribute-filter
-        configuration:
-          key: "test.different"
-          values:
-            - "key 1"
-            - "key 2"
-    numeric-attribute-filter4:
-        exporters:
-          - jaeger4
-          - jaeger5
-        policy: numeric-attribute-filter
-        configuration:
-          key: "http.status_code"
-          min-value: 400
-          max-value: 999
